After careful consideration we dedicided to NOT remove any triples
Publication is a formal step.
Deleting triples might have unwanted consequences.
Consider:
If there is already a rechtsgrond derived from the published besluit.
The published besluit gets removed. What next?

Of course, this might lead to weird UX for user.

The question is of course: where do we want our knowledge from?

So, we have no idea. And further optimisation of the flow is required.
TBD