Add SEARCH_GRAPHS_BLACKLIST
As a workaround. It seems that in some cases stuff is happening that
should not be creating a delta.

In this case, the gelinkt-notuleren-consumer service first imports
everything in a temporary graph, once done this is move to the final
graph. This entailed the following issues:

- the service gets notified too early, stuff happening in the
tempgraph should not trigger deltas.
- while the service is processing the data (from the temp-graph) the
source graph changes, and the current service is working on
inconsistent state.

Hence the workaround. (it is 8 in the morning on sunday, tomorrow is
demo)

There might be probably more elegant solutions:
- gelinkt-notuleren-consumer not triggering notifications. (but afik
not possible so far)
- make sure the service is working on the most actual version of the
data. (This might result in a lot of boilerplate to perform checks, I
don't think we've ever written a service like this)