Handle case of more than one treatment in a decision
A fixed bug involving both the prepublisher and templates means that
there is the possibility of a document being produced from an old
template but processed with the new prepublisher. This produces 2
triples with the `prov:generated` predicate on the Besluit. This change
handles that by ignoring the one that is not linked to an AP.